+++ changelog.md
@@ -1,3 +1,4 @@
+- add rough xml "extension" support to links
 - fix bug in arbitrary attributes that required there to be two or more
 
 # 0.7.0